# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  orchestrator.OptimizationConfig:
    properties:
      method:
        type: string
      parameters:
        additionalProperties: true
        type: object
    type: object
  orchestrator.OptimizationRequest:
    properties:
      grid_size:
        type: integer
      max_iterations:
        type: integer
      method:
        type: string
      optimization:
        $ref: '#/definitions/orchestrator.OptimizationConfig'
      parameters:
        additionalProperties: true
        type: object
      request_id:
        type: string
      strategy_id:
        type: string
      time_range:
        $ref: '#/definitions/orchestrator.TimeRange'
    type: object
  orchestrator.TimeRange:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Quantitative Contract Automated Trading System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QCAT API
  version: "1.0"
paths:
...
  /orchestrator/optimize:
    post:
      consumes:
      - application/json
      description: Submit an optimization request to the orchestrator
      parameters:
      - description: Optimization parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/orchestrator.OptimizationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              status:
                type: string
              task_id:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Submit optimization request
      tags:
      - Orchestrator
  /orchestrator/services:
    get:
      consumes:
      - application/json
      description: Get detailed information about all managed services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get services information
      tags:
      - Orchestrator
  /orchestrator/services/restart:
    post:
      consumes:
      - application/json
      description: Restart a specific managed service
      parameters:
      - description: Service to restart
        in: body
        name: request
        required: true
        schema:
          properties:
            service_name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Restart service
      tags:
      - Orchestrator
  /orchestrator/services/start:
    post:
      consumes:
      - application/json
      description: Start a specific managed service
      parameters:
      - description: Service to start
        in: body
        name: request
        required: true
        schema:
          properties:
            service_name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Start service
      tags:
      - Orchestrator
  /orchestrator/services/stop:
    post:
      consumes:
      - application/json
      description: Stop a specific managed service
      parameters:
      - description: Service to stop
        in: body
        name: request
        required: true
        schema:
          properties:
            service_name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Stop service
      tags:
      - Orchestrator
  /orchestrator/status:
    get:
      consumes:
      - application/json
      description: Get overall status of the orchestrator and all managed services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              services:
                type: object
              status:
                type: string
              timestamp:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get orchestrator status
      tags:
      - Orchestrator
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"